% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateIIF.R
\name{calculateIIF}
\alias{calculateIIF}
\title{Calculate Item Information Function}
\usage{
calculateIIF(
  theta,
  A = rep(1, length(B)),
  B = 0,
  C = rep(0, length(B)),
  D = 1.7
)
}
\arguments{
\item{theta}{Vector of time intensity parameters.}

\item{A}{Vector of discrimination parameters.}

\item{B}{Vector of difficulty parameters.}

\item{C}{Vector of guessing parameters.}

\item{D}{A constant.}
}
\value{
a matrix, with columns for different \code{theta} and rows for different items
}
\description{
Calculate item information function given item parameters of the 1PL, 2PL or 3PL IRT model.
}
