% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint-methods.R
\name{combineConstraints}
\alias{combineConstraints}
\title{Combine constraints}
\usage{
combineConstraints(..., message = TRUE)
}
\arguments{
\item{...}{multiple constraint-objects or a list with multiple constraint-objects}

\item{message}{A logical indicating whether a message should be given when only one constraint object is combined.}
}
\value{
A \code{data.frame} of block exclusions.
}
\description{
Combine multiple constraint-objects into one constraint object.
}
\examples{
combineConstraints(
 itemValuesConstraint(2, 1:10, operator = ">=", targetValue = 4),
 itemValuesConstraint(2, 1:10, operator = "<=", targetValue = 6)
)

}
