% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateExpectedRT.R
\name{calculateExpectedRT}
\alias{calculateExpectedRT}
\alias{calculateExpectedRTvar}
\title{Calculate Cumulants Lognormal Response Time Distribution}
\usage{
calculateExpectedRT(lambda, phi, zeta, sdEpsi)

calculateExpectedRTvar(lambda, phi, zeta, sdEpsi)
}
\arguments{
\item{lambda}{Vector of time intensity parameters.}

\item{phi}{[optional] Vector of speed sensitivity parameters.}

\item{zeta}{Vector of person speed parameters.}

\item{sdEpsi}{Vector of item specific residual variances.}
}
\description{
These functions have been deprecated. See \code{\link{getMean3PLN}} or \code{\link{getVar3PLN}} instead.
}
\section{Functions}{
\itemize{
\item \code{calculateExpectedRT()}: Calculate mean 3PLN

\item \code{calculateExpectedRTvar()}: Calculate mean 2PLN

}}
