% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gurobiExample.R
\docType{data}
\name{gurobiExample}
\alias{gurobiExample}
\title{Example \code{Gurobi} output.}
\format{
A \code{list}.
}
\usage{
gurobiExample
}
\description{
An example \code{Gurobi} output containing the optimal solution for the example in the vignette.
}
\keyword{datasets}
