% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectSolution.R
\name{inspectSolution}
\alias{inspectSolution}
\title{Inspect \code{useSolver} output}
\usage{
inspectSolution(solverOut, items, colNames, colSums = TRUE)
}
\arguments{
\item{solverOut}{Object created by \code{useSolver} function.}

\item{items}{Original \code{data.frame} containing information on item level.}

\item{colNames}{Which columns should be used from the \code{items} \code{data.frame}?}

\item{colSums}{Should column sums be calculated in the output? Only works if all columns are numeric.}
}
\value{
A \code{list} with assembled blocks as entries. Rows are the individual items. A final row is added, containing
the sums of each column.
}
\description{
Process a \code{useSolver} output of a successfully solved optimization problem to a list so it becomes humanly readable.
}
\examples{
## Example item pool
items <- data.frame(ID = 1:10,
itemValues = c(-4, -4, -2, -2, -1, -1, 20, 20, 0, 0))

## Test Assembly
usage <- itemUsageConstraint(nForms = 2, nItems = 10, operator = "=", targetValue = 1)
perForm <- itemsPerFormConstraint(nForms = 2, nItems = 10, operator = "=", targetValue = 5)
target <- itemTargetConstraint(nForms = 2, nItems = 10,
                               itemValues = items$itemValues,
                               targetValue = 0)
sol <- useSolver(allConstraints = list(usage, perForm, target),
                                  nForms = 2, nItems = 10, solver = "lpSolve")

## Inspect Solution
out <- inspectSolution(sol, items = items, colNames = "itemValues")

}
