% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemCategoryMinMax.R
\name{itemCategoryMinMax}
\alias{itemCategoryMinMax}
\alias{itemCategoryMin}
\alias{itemCategoryMax}
\alias{itemCategoryThreshold}
\title{Create item category constraints with minimum and maximum.}
\usage{
itemCategoryMinMax(nForms, nItems, itemCategories, min, max)

itemCategoryMin(nForms, nItems, itemCategories, min)

itemCategoryMax(nForms, nItems, itemCategories, max)

itemCategoryThreshold(nForms, nItems, itemCategories, targetValues, thresholds)
}
\arguments{
\item{nForms}{Number of forms to be created.}

\item{nItems}{Number of items in the item pool.}

\item{itemCategories}{a factor representing the categories/grouping of the items}

\item{min}{the minimal sum of the \code{itemValues} per test form}

\item{max}{the minimal sum of the \code{itemValues} per test form}

\item{targetValues}{an integer vector representing the target number per category.
The order of the target values should correspond with the order of the levels
of the factor in \code{itemCategory}.}

\item{thresholds}{an integer vector representing the thresholds per category.
The order of the thresholds should correspond with the order of the levels
of the factor in \code{itemCategories}.}
}
\value{
A sparse matrix.
}
\description{
Create constraints related to item categories/groupings (as
represented by \code{itemCategories}). That is, the created
constraints assure that the number of items of each category per test form is either
smaller or equal than the corresponding \code{max}, greater than or equal to \code{min} or both.
}
\details{
\code{itemCategoriesThreshold} also constrains the minimal and the maximal value
of the number of items of each category per test form, but based on chosen
\code{targetValues}, and maximal allowed deviations (i.e., \code{threshold})
from those \code{targetValues}.
}
\section{Functions}{
\itemize{
\item \code{itemCategoryMin}: constrain minimum value

\item \code{itemCategoryMax}: constrain maximum value

\item \code{itemCategoryThreshold}: constrain the distance form the \code{targetValues}
}}

\examples{
## constraints to make sure that there are at least 2 and maximally 4
##  items of each item type in each test form
nItems <- 30
item_type <- factor(sample(1:3, size = nItems, replace = TRUE))
itemCategoryMinMax(2, nItems, item_type, min = rep(2, 3), max = rep(4, 3))

## or alternatively
itemCategoryThreshold(2, nItems, item_type, targetValues = rep(3, 3), thresholds = rep(4, 3))

}
