% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemExclusionConstraint.R
\name{itemExclusionConstraint}
\alias{itemExclusionConstraint}
\title{Create item exclusion constraints.}
\usage{
itemExclusionConstraint(nForms, exclusionTuples, itemIDs)
}
\arguments{
\item{nForms}{Number of forms to be created.}

\item{exclusionTuples}{\code{data.frame} with two columns, containing tuples with item IDs which should be in test forms exclusively.}

\item{itemIDs}{Character vector of item IDs in correct order.}
}
\value{
A sparse matrix.
}
\description{
tbd
}
\examples{
#tbd

}
