% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemExclusionTransform.R
\name{itemExclusionTuples}
\alias{itemExclusionTuples}
\title{Create item exclusion tuples from Item DB output.}
\usage{
itemExclusionTuples(items, idCol = "ID", exclusions)
}
\arguments{
\item{items}{\code{data.frame} as exported by \code{Item DB}.}

\item{idCol}{Name of the item ID column.}

\item{exclusions}{Name of the exclusion column.}
}
\value{
A sparse matrix.
}
\description{
tbd
}
\examples{
#tbd

}
