% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemPoolConstraint.R
\name{itemPoolConstraint}
\alias{itemPoolConstraint}
\title{Use complete item pool.}
\usage{
itemPoolConstraint(nForms, nItems)
}
\arguments{
\item{nForms}{Number of forms to be created.}

\item{nItems}{Number of items in the item pool.}
}
\value{
A sparse matrix.
}
\description{
Currently not in use as noItemOverlapConstraint does this. Development incomplete!
}
\examples{
#tbd

}
