% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items_pools.R
\docType{data}
\name{items_diao}
\alias{items_diao}
\title{Small simulated item pool example.}
\format{
A \code{data.frame} .
\describe{
\item{item}{Item identifier.}
\item{a}{Discrimination parameter.}
\item{b}{Difficulty parameter.}
\item{c}{Pseudo-guessing parameter.}
\item{category}{Content category.}
}
}
\usage{
items_diao
}
\description{
A \code{data.frame} containing 165 items calibrated using a 3PL model. This item pool is analogous to one of the item pools used
in Diao & van der Linden (2011).
}
\references{
Diao, Q. & van der Linden, W.J. (2011). Automated test assembly using lp_solve version 5.5 in R. \emph{Applied Psychological Measurement, 35 (5)}, 398-409.
}
\keyword{datasets}
