% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items_pools.R
\docType{data}
\name{items_lsa}
\alias{items_lsa}
\title{Simulated item pool example.}
\format{
A \code{data.frame} .
\describe{
\item{testlet}{Testlet identifier (items in the same testlet share a common stimulus.}
\item{item}{Item identifier.}
\item{level}{Competence level.}
\item{format}{Item format.}
\item{frequency}{Solution frequency.}
\item{infit}{Item infit.}
\item{time}{Average response time in seconds.}
\item{anchor}{Is the item an anchor item?}
}
}
\usage{
items_lsa
}
\description{
A \code{data.frame} containing 209 calibrated items with different categorical and metric properties, comparable to an item pool from a large-scale
assessment.
}
\keyword{datasets}
