% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items_pools.R
\docType{data}
\name{items_mini}
\alias{items_mini}
\title{Small simulated item pool example.}
\format{
A \code{data.frame} .
\describe{
\item{item}{Item identifier.}
\item{format}{Item format (e.g., multiple choice, open answer, order item).}
\item{time}{Average response time in seconds.}
\item{difficulty}{IRT difficulty parameter.}
}
}
\usage{
items_mini
}
\description{
A \code{data.frame} containing 30 items with different categorical and metric properties.
}
\keyword{datasets}
