% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items_pools.R
\docType{data}
\name{items_pilot}
\alias{items_pilot}
\title{Small simulated item pool example.}
\format{
A \code{data.frame} .
\describe{
\item{item}{Item identifier.}
\item{diffCategory}{Item difficulty (five categories).}
\item{format}{Item format (multiple choice, constructed multiple choice, or open answer).}
\item{domain}{Item domain (listening, reading, or writing).}
\item{time}{Average response times in seconds.}
\item{exclusions}{Items which can not be in the same test form.}
}
}
\usage{
items_pilot
}
\description{
A \code{data.frame} containing 100 not yet calibrated items with different categorical and metric properties.
}
\keyword{datasets}
