% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items_sim.R
\docType{data}
\name{items_sim}
\alias{items_sim}
\title{Small simulated item pool example.}
\format{
A \code{data.frame} .
\describe{
\item{id}{Item identifier.}
\item{format}{Items format (multiple choice, open answer, ordering).}
\item{mean_time}{Average response times in seconds.}
\item{difficulty}{IRT difficulty parameter.}
}
}
\usage{
items_sim
}
\description{
A \code{data.frame} containing 30 items with different categorical and metric properties.
}
\keyword{datasets}
