% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemTargetConstraint.R
\name{itemTargetConstraint}
\alias{itemTargetConstraint}
\title{Create optimization constraints.}
\usage{
itemTargetConstraint(nForms, nItems, itemValues, targetValue)
}
\arguments{
\item{nForms}{Number of forms to be created.}

\item{nItems}{Number of items in the item pool.}

\item{itemValues}{Item parameter/values for which the sum per test form should be constrained}

\item{targetValue}{The target value to be used in the constraints.}
}
\value{
A sparse matrix.
}
\description{
Create constraints that define the optimization goal of a automated test assembly problem.
}
\examples{
#tbd

}
