% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemValuesMinMax.R
\name{itemValuesMinMax}
\alias{itemValuesMinMax}
\alias{itemValuesMin}
\alias{itemValuesMax}
\alias{itemValuesThreshold}
\title{Create single value constraints with minimum and maximum.}
\usage{
itemValuesMinMax(nForms, nItems, itemValues, min, max)

itemValuesMin(nForms, nItems, itemValues, min)

itemValuesMax(nForms, nItems, itemValues, max)

itemValuesThreshold(nForms, nItems, itemValues, targetValue, threshold)
}
\arguments{
\item{nForms}{Number of forms to be created.}

\item{nItems}{Number of items in the item pool.}

\item{itemValues}{Item parameter/values for which the sum per test form should be constrained.}

\item{min}{the minimal sum of the \code{itemValues} per test form}

\item{max}{the minimal sum of the \code{itemValues} per test form}

\item{targetValue}{the target test form value}

\item{threshold}{the maximum allowed difference from the \code{targetValue}}
}
\value{
A sparse matrix.
}
\description{
Create constraints related to an item parameter/value. That is, the created
constraints assure that the sum of the \code{itemValues} is smaller than or equal
to \code{max}, greater dan or equal to \code{min}, or both.
}
\details{
\code{itemValuesThreshold} also constrains the minimal and the maximal value
of the sum of the \code{itemValues}, but based on a chosen
and a maximal allowed deviation (i.e., \code{threshold}) from that \code{targetValue}.
}
\section{Functions}{
\itemize{
\item \code{itemValuesMin}: constrain minimum value

\item \code{itemValuesMax}: constrain maximum value

\item \code{itemValuesThreshold}: constrain the distance form the targetValue
}}

\examples{
## constraints to make sure that the sum of the item values (1:10) is between
## 4 and 6
itemValuesMinMax (2, 10, 1:10, min = 4, max = 6)

## or alternatively
itemValuesThreshold (2, 10, 1:10, targetValue = 5, threshold = 1)

}
