% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectiveConstraint.R
\name{maxConstraint}
\alias{maxConstraint}
\title{Max Constraint.}
\usage{
maxConstraint(
  nItems,
  itemValues,
  info_text = NULL,
  itemIDs = names(itemValues)
)
}
\arguments{
\item{nItems}{Number of items in the item pool.}

\item{itemValues}{Item parameter/values for which the sum per test form should be constrained.}

\item{info_text}{a character string of length 1, to be used in the \code{"info"}-attribute of the resulting \code{constraint}-object.}

\item{itemIDs}{a character vector of item IDs in correct ordering, or NULL.}
}
\value{
An object of class \code{"constraint"}.
}
\description{
Create max-constraints related to an item parameter/value. That is, the created
constraints can be used to maximize the sum of the
item values (\code{itemValues}) of the test form.
Note that this constraint can only be used when only one test form has to be assembled.
}
\examples{
# constraint that maximizes the sum of the itemValues
maxConstraint(nItems = 10,itemValues = rep(-2:2, 2))

}
