% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noItemOverlapConstraint.R
\name{noItemOverlapConstraint}
\alias{noItemOverlapConstraint}
\title{Create no item overlap constraints.}
\usage{
noItemOverlapConstraint(nForms, nItems, sign)
}
\arguments{
\item{nForms}{Number of forms to be created.}

\item{nItems}{Number of items in the item pool.}

\item{sign}{tbd}
}
\value{
A sparse matrix.
}
\description{
tbd
}
\examples{
#tbd

}
