% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrItInFormConstraint.R
\name{nrItInFormConstraint}
\alias{nrItInFormConstraint}
\title{Create nomber of items per test form constraints.}
\usage{
nrItInFormConstraint(nForms, nItems, sign, nrItems)
}
\arguments{
\item{nForms}{Number of forms to be created.}

\item{nItems}{Number of items in the item pool.}

\item{sign}{tbd}

\item{nrItems}{Fixed number of items per test form.}
}
\value{
A sparse matrix.
}
\description{
tbd
}
\examples{
#tbd

}
