% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGurobiOutput.R
\name{processGurobiOutput}
\alias{processGurobiOutput}
\title{Process gurobi output for readability.}
\usage{
processGurobiOutput(gurobiObj, items, nForms, nItems, output = "data.frame")
}
\arguments{
\item{gurobiObj}{Object created by \code{Gurobi} solver.}

\item{items}{Original \code{data.frame} containing information on item level.}

\item{nForms}{Number of forms to be created.}

\item{nItems}{Number of items in the item pool.}

\item{output}{Should the output be returned as a \code{data.frame} or a \code{list}.}
}
\value{
a \code{data.frame} or a \code{list}.
}
\description{
tbd
}
\examples{
#tbd

}
