% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetConstraint.R
\name{targetConstraint}
\alias{targetConstraint}
\title{Create optimization constraints.}
\usage{
targetConstraint(
  nForms,
  nItems,
  itemValues,
  targetValues,
  thetaPoints,
  relative = TRUE,
  bWidth = 0
)
}
\arguments{
\item{nForms}{Number of forms to be created.}

\item{nItems}{Number of items in the item pool.}

\item{itemValues}{tbd}

\item{targetValues}{tbd}

\item{thetaPoints}{tbd}

\item{relative}{tbd}

\item{bWidth}{tbd}
}
\value{
A sparse matrix.
}
\description{
tbd
}
\examples{
#tbd

}
