% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateExpectedRT.R
\name{calculateExpectedRT}
\alias{calculateExpectedRT}
\title{Calculate Expected Response Times}
\usage{
calculateExpectedRT(lambda, phi, zeta, sdEpsi)
}
\arguments{
\item{lambda}{Vector of time intensity parameters.}

\item{phi}{Vector of time sensitivity parameters.}

\item{zeta}{Vector of person speed parameters.}

\item{sdEpsi}{Vector of item specific residual variances.}
}
\value{
a vector
}
\description{
Calculate expected response times given item parameters of the log normal response time model.
}
\details{
Expected response times are calculated according to the log normal response time models by
van der Linden (2006) or Klein Entink et al. (2009). If \code{phi} is \code{1}, the model
by van der Linden (2006) is used. Either a single set of parameters of vectors of each parameters
can be supplied. If multiple \code{zeta} are supplied, a list
}
