% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noItemOverlapConstraint.R
\name{noItemOverlapConstraint}
\alias{noItemOverlapConstraint}
\title{Create no item overlap constraints.}
\usage{
noItemOverlapConstraint(nForms, nItems, sign = -1)
}
\arguments{
\item{nForms}{Number of forms to be created.}

\item{nItems}{Number of items in the item pool.}

\item{sign}{tbd}
}
\value{
A sparse matrix.
}
\description{
Creates constraints to assure no item overlap. If \code{sign} is set to \code{0}, complete item pool depletion is also assured.
}
\examples{
#tbd

}
