% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareConstraints.R
\name{prepareConstraints}
\alias{prepareConstraints}
\title{Prepare a list of constraints for \code{Gurobi}.}
\usage{
prepareConstraints(allConstraints, M, n_items, f)
}
\arguments{
\item{allConstraints}{Number of forms to be created.}

\item{M}{tbd}

\item{n_items}{tbd}

\item{f}{tbd}
}
\value{
tbd
}
\description{
tbd
}
\examples{
#tbd

}
