% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleParameterConstraint.R
\name{singleParameterConstraint}
\alias{singleParameterConstraint}
\title{Create single value constraints with a tolerance.}
\usage{
singleParameterConstraint(
  nForms,
  nItems,
  itemValues,
  targetValues = NULL,
  tolerance
)
}
\arguments{
\item{nForms}{Number of forms to be created.}

\item{nItems}{Number of items in the item pool.}

\item{itemValues}{item parameter/values for which the sum per test form should be constrained}

\item{targetValues}{the target test form value}

\item{tolerance}{tbd}
}
\value{
A sparse matrix.
}
\description{
tbd
}
\examples{
#tbd

}
